fix: increase metric precision

+++ cycling_utils/metrics.py
@@ -63,7 +63,7 @@ class MetricsTracker:
     def reduce(self):
         names, local = zip(*self.local.items())
         local = torch.tensor(
-            local, dtype=torch.float16, requires_grad=False, device="cuda"
+            local, dtype=torch.float32, requires_grad=False, device="cuda"
         )
         dist.all_reduce(local, op=dist.ReduceOp.SUM)
         self.local = defaultdict(float, zip(names, local.tolist()))
